requirements.txt
-i https://pypi.org/simple
aenum==3.1.15
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.11.7; python_version >= '3.9'
aiosignal==1.3.1; python_version >= '3.7'
aniso8601==9.0.1
annotated-types==0.7.0; python_version >= '3.8'
attrs==24.2.0; python_version >= '3.7'
bcrypt==4.2.1; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
datetime==5.5; python_version >= '3.7'
deprecated==1.2.15; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dnspython==2.7.0; python_version >= '3.9'
flask==3.1.0; python_version >= '3.9'
flask-restful==0.3.10
frozenlist==1.5.0; python_version >= '3.8'
future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
google-api-core==2.23.0; python_version >= '3.7'
google-api-python-client==2.154.0; python_version >= '3.7'
google-auth==2.36.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1; python_version >= '3.6'
google-oauth2-tool==0.0.3; python_version >= '3.5'
googleapis-common-protos==1.66.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
line-bot-sdk==3.14.2; python_full_version >= '3.9.0'
markupsafe==3.0.2; python_version >= '3.9'
multidict==6.1.0; python_version >= '3.8'
oauth2client==4.1.3
oauthlib==3.2.2; python_version >= '3.6'
pillow==11.0.0; python_version >= '3.9'
propcache==0.2.0; python_version >= '3.8'
proto-plus==1.25.0; python_version >= '3.7'
protobuf==5.28.3; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pydantic==2.10.1; python_version >= '3.8'
pydantic-core==2.27.1; python_version >= '3.8'
pymongo==4.10.1; python_version >= '3.8'
pyparsing==3.2.0; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
qrcode==8.0; python_version >= '3.9' and python_version < '4.0'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==75.6.0; python_version >= '3.9'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.12.2; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.3; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
wrapt==1.17.0; python_version >= '3.8'
yarl==1.18.0; python_version >= '3.9'
zope.interface==7.1.1; python_version >= '3.8'
